method for inventory return implemented
