+++ src/main/java/ua/javarush/converter/YearConverter.java
@@ -1,0 +1,25 @@
+package ua.javarush.converter;
+
+import jakarta.persistence.AttributeConverter;
+import jakarta.persistence.Converter;
+
+import java.time.Year;
+
+@Converter(autoApply = true)
+public class YearConverter implements AttributeConverter<Year, Short> {
+    @Override
+    public Short convertToDatabaseColumn(Year year) {
+        if (year != null) {
+            return (short) year.getValue();
+        }
+        return null;
+    }
+
+    @Override
+    public Year convertToEntityAttribute(Short value) {
+        if (value != null) {
+            return Year.of(value);
+        }
+        return null;
+    }
+}
