+++ src/main/java/ua/javarush/dao/AddressDAO.java
@@ -1,7 +1,7 @@
 package ua.javarush.dao;
 
 import org.hibernate.SessionFactory;
-import ua.javarush.entities.Address;
+import ua.javarush.entity.Address;
 
 public class AddressDAO extends AbstractDAO<Address> {
