+++ src/main/java/ua/javarush/dao/CategoryDAO.java
@@ -1,7 +1,7 @@
 package ua.javarush.dao;
 
 import org.hibernate.SessionFactory;
-import ua.javarush.entities.Category;
+import ua.javarush.entity.Category;
 
 public class CategoryDAO extends AbstractDAO<Category> {
