+++ src/main/java/ua/javarush/dao/FilmTextDAO.java
@@ -1,7 +1,7 @@
 package ua.javarush.dao;
 
 import org.hibernate.SessionFactory;
-import ua.javarush.entities.FilmText;
+import ua.javarush.entity.FilmText;
 
 public class FilmTextDAO extends AbstractDAO<FilmText> {
