+++ src/main/java/ua/javarush/dao/StaffDAO.java
@@ -1,7 +1,7 @@
 package ua.javarush.dao;
 
 import org.hibernate.SessionFactory;
-import ua.javarush.entities.Staff;
+import ua.javarush.entity.Staff;
 
 public class StaffDAO extends AbstractDAO<Staff> {
