--- src/main/java/ua/javarush/entities/Staff.java
@@ -1,61 +1,0 @@
-package ua.javarush.entities;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.Lob;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import lombok.Data;
-import org.hibernate.annotations.Type;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import java.time.LocalDateTime;
-
-@Data
-@Entity
-@Table(name = "staff", schema = "movie")
-public class Staff {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "staff_id")
-    private Byte id;
-
-    @Column(name = "first_name")
-    private String firstname;
-
-    @Column(name = "last_name")
-    private String lastName;
-
-    @ManyToOne
-    @JoinColumn(name = "address_id")
-    private Address address;
-
-    @Lob
-    @Column(name = "picture", columnDefinition = "blob")
-    private Byte[] picture;
-
-    @Column(name = "email")
-    private String email;
-
-    @ManyToOne
-    @JoinColumn(name = "store_id")
-    private Store store;
-
-    @Column(name = "active", columnDefinition = "bit")
-    @Type(type = "org.hibernate.type.NumericBooleanType")
-    private Boolean isActive;
-
-    @Column(name = "username")
-    private String username;
-
-    @Column(name = "password")
-    private String password;
-
-    @UpdateTimestamp
-    @Column(name = "last_update")
-    private LocalDateTime lastUpdate;
-}
