+++ src/main/java/ua/javarush/entity/Address.java
@@ -1,0 +1,47 @@
+package ua.javarush.entity;
+
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
+import lombok.Data;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.LocalDateTime;
+
+@Data
+@Entity
+@Table(name = "address", schema = "movie")
+public class Address {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "address_id")
+    private Short id;
+
+    @Column(name = "address")
+    private String address;
+
+    @Column(name = "address2")
+    private String address2;
+
+    @Column(name = "district")
+    private String district;
+
+    @ManyToOne
+    @JoinColumn(name = "city_id")
+    private City city;
+
+    @Column(name = "postal_code")
+    private String postalCode;
+
+    @Column(name = "phone")
+    private String phone;
+
+    @UpdateTimestamp
+    @Column(name = "last_update")
+    private LocalDateTime lastUpdate;
+}
