+++ src/main/java/ua/javarush/entity/City.java
@@ -1,10 +1,12 @@
-package ua.javarush.entities;
+package ua.javarush.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.JoinColumn;
+import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Data;
 import org.hibernate.annotations.UpdateTimestamp;
@@ -13,15 +15,19 @@ import java.time.LocalDateTime;
 
 @Data
 @Entity
-@Table(name = "country", schema = "movie")
-public class Country {
+@Table(name = "city", schema = "movie")
+public class City {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "country_id")
+    @Column(name = "city_id")
     private Short id;
 
-    @Column(name = "country")
+    @Column(name = "city")
     private String name;
+
+    @ManyToOne
+    @JoinColumn(name = "country_id")
+    private Country country;
 
     @UpdateTimestamp
     @Column(name = "last_update")
