+++ src/main/java/ua/javarush/entity/Country.java
@@ -1,12 +1,10 @@
-package ua.javarush.entities;
+package ua.javarush.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import lombok.Data;
 import org.hibernate.annotations.UpdateTimestamp;
@@ -15,20 +13,15 @@ import java.time.LocalDateTime;
 
 @Data
 @Entity
-@Table(name = "inventory", schema = "movie")
-public class Inventory {
+@Table(name = "country", schema = "movie")
+public class Country {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "inventory_id")
-    private Integer id;
+    @Column(name = "country_id")
+    private Short id;
 
-    @ManyToOne
-    @JoinColumn(name = "film_id")
-    private Film film;
-
-    @ManyToOne
-    @JoinColumn(name = "store_id")
-    private Store store;
+    @Column(name = "country")
+    private String name;
 
     @UpdateTimestamp
     @Column(name = "last_update")
