+++ src/main/java/ua/javarush/entity/Customer.java
@@ -1,4 +1,4 @@
-package ua.javarush.entities;
+package ua.javarush.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -11,38 +11,44 @@ import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Data;
 import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.Type;
 import org.hibernate.annotations.UpdateTimestamp;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 @Entity
-@Table(name = "payment", schema = "movie")
-public class Payment {
+@Table(name = "customer", schema = "movie")
+public class Customer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "payment_id")
+    @Column(name = "customer_id")
     private Short id;
 
     @ManyToOne
-    @JoinColumn(name = "customer_id")
-    private Customer customer;
+    @JoinColumn(name = "store_id")
+    private Store store;
 
-    @ManyToOne
-    @JoinColumn(name = "staff_id")
-    private Staff staff;
+    @Column(name = "first_name")
+    private String firstname;
+
+    @Column(name = "last_name")
+    private String lastName;
+
+    @Column(name = "email")
+    private String email;
 
     @OneToOne
-    @JoinColumn(name = "rental_id")
-    private Rental rental;
+    @JoinColumn(name = "address_id")
+    private Address address;
 
-    @Column(name = "amount")
-    private BigDecimal amount;
+    @Column(name = "active", columnDefinition = "bit")
+    @Type(type = "org.hibernate.type.NumericBooleanType")
+    private Boolean isActive;
 
     @CreationTimestamp
-    @Column(name = "payment_date")
-    private LocalDateTime paymentDate;
+    @Column(name = "create_date")
+    private LocalDateTime createDate;
 
     @UpdateTimestamp
     @Column(name = "last_update")
