+++ src/main/java/ua/javarush/entity/Language.java
@@ -1,12 +1,10 @@
-package ua.javarush.entities;
+package ua.javarush.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Data;
 import org.hibernate.annotations.UpdateTimestamp;
@@ -15,20 +13,15 @@ import java.time.LocalDateTime;
 
 @Data
 @Entity
-@Table(name = "store", schema = "movie")
-public class Store {
+@Table(name = "language", schema = "movie")
+public class Language {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "store_id")
+    @Column(name = "language_id")
     private Byte id;
 
-    @OneToOne
-    @JoinColumn(name = "manager_staff_id")
-    private Staff manager;
-
-    @OneToOne
-    @JoinColumn(name = "address_id")
-    private Address address;
+    @Column(name = "name", columnDefinition = "char")
+    private String name;
 
     @UpdateTimestamp
     @Column(name = "last_update")
