+++ src/main/java/ua/javarush/entity/Payment.java
@@ -1,4 +1,4 @@
-package ua.javarush.entities;
+package ua.javarush.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -7,38 +7,42 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Data;
+import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
 @Entity
-@Table(name = "rental", schema = "movie")
-public class Rental {
+@Table(name = "payment", schema = "movie")
+public class Payment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "rental_id")
-    private Integer id;
-
-    @Column(name = "rental_date")
-    private LocalDateTime rentalDate;
-
-    @ManyToOne
-    @JoinColumn(name = "inventory_id")
-    private Inventory inventory;
+    @Column(name = "payment_id")
+    private Short id;
 
     @ManyToOne
     @JoinColumn(name = "customer_id")
     private Customer customer;
 
-    @Column(name = "return_date")
-    private LocalDateTime returnDate;
-
     @ManyToOne
     @JoinColumn(name = "staff_id")
     private Staff staff;
+
+    @OneToOne
+    @JoinColumn(name = "rental_id")
+    private Rental rental;
+
+    @Column(name = "amount")
+    private BigDecimal amount;
+
+    @CreationTimestamp
+    @Column(name = "payment_date")
+    private LocalDateTime paymentDate;
 
     @UpdateTimestamp
     @Column(name = "last_update")
