+++ src/main/java/ua/javarush/entity/Rental.java
@@ -1,4 +1,4 @@
-package ua.javarush.entities;
+package ua.javarush.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -8,7 +8,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.hibernate.annotations.UpdateTimestamp;
 
@@ -16,31 +15,30 @@ import java.time.LocalDateTime;
 
 @Data
 @Entity
-@Table(name = "address", schema = "movie")
-public class Address {
+@Table(name = "rental", schema = "movie")
+public class Rental {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "address_id")
-    private Short id;
+    @Column(name = "rental_id")
+    private Integer id;
 
-    @Column(name = "address")
-    private String address;
-
-    @Column(name = "address2")
-    private String address2;
-
-    @Column(name = "district")
-    private String district;
+    @Column(name = "rental_date")
+    private LocalDateTime rentalDate;
 
     @ManyToOne
-    @JoinColumn(name = "city_id")
-    private City city;
+    @JoinColumn(name = "inventory_id")
+    private Inventory inventory;
 
-    @Column(name = "postal_code")
-    private String postalCode;
+    @ManyToOne
+    @JoinColumn(name = "customer_id")
+    private Customer customer;
 
-    @Column(name = "phone")
-    private String phone;
+    @Column(name = "return_date")
+    private LocalDateTime returnDate;
+
+    @ManyToOne
+    @JoinColumn(name = "staff_id")
+    private Staff staff;
 
     @UpdateTimestamp
     @Column(name = "last_update")
