+++ src/main/java/ua/javarush/entity/Staff.java
@@ -1,4 +1,4 @@
-package ua.javarush.entities;
+package ua.javarush.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -6,11 +6,10 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
+import jakarta.persistence.Lob;
 import jakarta.persistence.ManyToOne;
-import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Data;
-import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.Type;
 import org.hibernate.annotations.UpdateTimestamp;
 
@@ -18,16 +17,12 @@ import java.time.LocalDateTime;
 
 @Data
 @Entity
-@Table(name = "customer", schema = "movie")
-public class Customer {
+@Table(name = "staff", schema = "movie")
+public class Staff {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "customer_id")
-    private Short id;
-
-    @ManyToOne
-    @JoinColumn(name = "store_id")
-    private Store store;
+    @Column(name = "staff_id")
+    private Byte id;
 
     @Column(name = "first_name")
     private String firstname;
@@ -35,20 +30,30 @@ public class Customer {
     @Column(name = "last_name")
     private String lastName;
 
+    @ManyToOne
+    @JoinColumn(name = "address_id")
+    private Address address;
+
+    @Lob
+    @Column(name = "picture", columnDefinition = "blob")
+    private Byte[] picture;
+
     @Column(name = "email")
     private String email;
 
-    @OneToOne
-    @JoinColumn(name = "address_id")
-    private Address address;
+    @ManyToOne
+    @JoinColumn(name = "store_id")
+    private Store store;
 
     @Column(name = "active", columnDefinition = "bit")
     @Type(type = "org.hibernate.type.NumericBooleanType")
     private Boolean isActive;
 
-    @CreationTimestamp
-    @Column(name = "create_date")
-    private LocalDateTime createDate;
+    @Column(name = "username")
+    private String username;
+
+    @Column(name = "password")
+    private String password;
 
     @UpdateTimestamp
     @Column(name = "last_update")
