+++ src/main/java/ua/javarush/entity/Store.java
@@ -1,4 +1,4 @@
-package ua.javarush.entities;
+package ua.javarush.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -6,7 +6,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
 import lombok.Data;
 import org.hibernate.annotations.UpdateTimestamp;
@@ -15,19 +15,20 @@ import java.time.LocalDateTime;
 
 @Data
 @Entity
-@Table(name = "city", schema = "movie")
-public class City {
+@Table(name = "store", schema = "movie")
+public class Store {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "city_id")
-    private Short id;
+    @Column(name = "store_id")
+    private Byte id;
 
-    @Column(name = "city")
-    private String name;
+    @OneToOne
+    @JoinColumn(name = "manager_staff_id")
+    private Staff manager;
 
-    @ManyToOne
-    @JoinColumn(name = "country_id")
-    private Country country;
+    @OneToOne
+    @JoinColumn(name = "address_id")
+    private Address address;
 
     @UpdateTimestamp
     @Column(name = "last_update")
